% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict1toMaxNeighbors.R
\name{knn}
\alias{knn}
\title{K nearest neighbors alogrithm}
\usage{
knn(x.mat, y.vec, testx.vec, max.neighbors)
}
\arguments{
\item{x.mat}{\link{n x p} matrix of features}

\item{y.vec}{label column vector \link{n}}

\item{testx.vec}{numeric feature vector for test \link{p}}

\item{max.neighbors}{max number of neighbors}
}
\value{
numeric vector of size max.neighbors with predicitons from one to max.neighbors
}
\description{
A package that computes the k nearest neighbors
This funciton calls the c interface in the package to calculate the k nearest neighbors for a given set of data
}
